fix: ios 와 web 쿠키 domain 이 달라서 controller 분리
