+++ src/main/java/sudols/ecopercent/controller/OAuth2Controller.java
@@ -10,25 +10,33 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import sudols.ecopercent.dto.oauth2.apple.AppleAuthorizationResponse;
-import sudols.ecopercent.service.oauth2.AppleOAuth2IosService;
-import sudols.ecopercent.service.oauth2.AppleOAuth2WebService;
-import sudols.ecopercent.service.oauth2.KakaoOAuth2Service;
+import sudols.ecopercent.service.AppleOAuth2IosService;
+import sudols.ecopercent.service.AppleOAuth2WebService;
+import sudols.ecopercent.service.KakaoOAuth2IosService;
+import sudols.ecopercent.service.KakaoOAuth2WebService;
 
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("/login/oauth2")
 public class OAuth2Controller {
 
-    private final KakaoOAuth2Service kakaoOAuth2Service;
+    private final KakaoOAuth2WebService kakaoOAuth2WebService;
+    private final KakaoOAuth2IosService kakaoOAuth2IosService;
     private final AppleOAuth2IosService appleOAuth2IosService;
     private final AppleOAuth2WebService appleOAuth2WebService;
 
-    @PostMapping("/kakao")
+    @PostMapping("/kakao/web")
     @ResponseBody
-    public ResponseEntity<?> kakaoOAuth2Login(HttpServletRequest request,
+    public ResponseEntity<?> kakaoOAuth2WebLogin(HttpServletRequest request,
                                               HttpServletResponse response) {
-        System.out.println("kakaoOAuth2Login");
-        return kakaoOAuth2Service.login(request, response);
+        return kakaoOAuth2WebService.login(request, response);
+    }
+
+    @PostMapping("/kakao/ios")
+    @ResponseBody
+    public ResponseEntity<?> kakaoOAuth2IosLogin(HttpServletRequest request,
+                                              HttpServletResponse response) {
+        return kakaoOAuth2IosService.login(request, response);
     }
 
     @PostMapping("/apple/ios")
