+++ src/main/java/sudols/ecopercent/service/AppleOAuth2IosService.java
@@ -1,4 +1,4 @@
-package sudols.ecopercent.service.oauth2;
+package sudols.ecopercent.service;
 
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -8,11 +8,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import sudols.ecopercent.domain.User;
 import sudols.ecopercent.dto.oauth2.SignupResponse;
-import sudols.ecopercent.dto.oauth2.apple.AppleTokenResponse;
+import sudols.ecopercent.dto.oauth2.apple.AppleSignInResponse;
 import sudols.ecopercent.repository.UserRepository;
-import sudols.ecopercent.security.AppleOAuth2Provider;
 import sudols.ecopercent.security.JwtTokenProvider;
-import sudols.ecopercent.security.OAuth2ResponseProvider;
+import sudols.ecopercent.service.provider.AppleOAuth2Provider;
 
 import java.security.PublicKey;
 import java.util.Optional;
@@ -24,7 +23,7 @@ public class AppleOAuth2IosService {
 
     private final UserRepository userRepository;
     private final JwtTokenProvider jwtTokenProvider;
-    private final OAuth2ResponseProvider oAuth2ResponseProvider;
+    private final sudols.ecopercent.service.provider.OAuth2ResponseProvider OAuth2ResponseProvider;
     private final AppleOAuth2Provider appleOAuth2Provider;
 
     public ResponseEntity<?> login(HttpServletRequest request) {
@@ -33,10 +32,10 @@ public class AppleOAuth2IosService {
         String email = jwtTokenProvider.getEmailFromTokenWithPublicKey(identityToken, publicKey);
         Optional<User> optionalUser = userRepository.findByEmail(email);
         if (optionalUser.isEmpty()) {
-            SignupResponse signupResponse = oAuth2ResponseProvider.generateAccessTokenAndGetSignupResponse(email);
+            SignupResponse signupResponse = OAuth2ResponseProvider.generateAccessTokenAndGetSignupResponse(email);
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(signupResponse);
         }
-        AppleTokenResponse appleTokenResponse = oAuth2ResponseProvider.generateTokenAndGetTokenResponse(optionalUser.get());
-        return ResponseEntity.status(HttpStatus.OK).body(appleTokenResponse);
+        AppleSignInResponse appleSignInResponse = OAuth2ResponseProvider.generateTokenAndGetTokenResponse(optionalUser.get());
+        return ResponseEntity.status(HttpStatus.OK).body(appleSignInResponse);
     }
 }
