+++ src/main/java/sudols/ecopercent/service/provider/OAuth2ResponseProvider.java
@@ -1,4 +1,4 @@
-package sudols.ecopercent.security;
+package sudols.ecopercent.service.provider;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,7 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import sudols.ecopercent.domain.User;
 import sudols.ecopercent.dto.oauth2.SignupResponse;
-import sudols.ecopercent.dto.oauth2.apple.AppleTokenResponse;
+import sudols.ecopercent.dto.oauth2.apple.AppleSignInResponse;
+import sudols.ecopercent.security.JwtTokenProvider;
 
 @Component
 @RequiredArgsConstructor
@@ -14,25 +15,48 @@ public class OAuth2ResponseProvider {
 
     private final JwtTokenProvider jwtTokenProvider;
 
-    public void generateAccessTokenAndAddCookie(HttpServletResponse response, String email, String domain) {
+    public void generateAccessTokenAndAddCookieForWeb(HttpServletResponse response, String email, String domain) {
         String access = jwtTokenProvider.generateAccessToken(email);
         Cookie accessTokenCookie = new Cookie("access", access);
-//        accessTokenCookie.setDomain(domain);
+        accessTokenCookie.setDomain(domain);
         accessTokenCookie.setPath("/");
         accessTokenCookie.setSecure(true);
         response.addCookie(accessTokenCookie);
     }
 
-    public void generateAccessRefreshTokenAndAddTokenCookie(HttpServletResponse response, User user, String domain) {
+    public void generateTokensAndAddCookieForWeb(HttpServletResponse response, User user, String domain) {
         String accessToken = jwtTokenProvider.generateAccessToken(user.getEmail());
         String refreshToken = jwtTokenProvider.generateRefreshToken(user.getEmail());
 
         Cookie accessTokenCookie = new Cookie("access", accessToken);
-//        accessTokenCookie.setDomain(domain);
+        accessTokenCookie.setDomain(domain);
         accessTokenCookie.setPath("/");
 
         Cookie refreshTokenCookie = new Cookie("refresh", refreshToken);
-//        refreshTokenCookie.setDomain(domain);
+        refreshTokenCookie.setDomain(domain);
+        refreshTokenCookie.setHttpOnly(true);
+        refreshTokenCookie.setPath("/");
+
+        response.addCookie(accessTokenCookie);
+        response.addCookie(refreshTokenCookie);
+    }
+
+    public void generateAccessTokenAndAddCookieForIOS(HttpServletResponse response, String email) {
+        String access = jwtTokenProvider.generateAccessToken(email);
+        Cookie accessTokenCookie = new Cookie("access", access);
+        accessTokenCookie.setPath("/");
+        accessTokenCookie.setSecure(true);
+        response.addCookie(accessTokenCookie);
+    }
+
+    public void generateTokensAndAddCookieForIos(HttpServletResponse response, User user) {
+        String accessToken = jwtTokenProvider.generateAccessToken(user.getEmail());
+        String refreshToken = jwtTokenProvider.generateRefreshToken(user.getEmail());
+
+        Cookie accessTokenCookie = new Cookie("access", accessToken);
+        accessTokenCookie.setPath("/");
+
+        Cookie refreshTokenCookie = new Cookie("refresh", refreshToken);
         refreshTokenCookie.setHttpOnly(true);
         refreshTokenCookie.setPath("/");
 
@@ -47,10 +71,10 @@ public class OAuth2ResponseProvider {
                 .build();
     }
 
-    public AppleTokenResponse generateTokenAndGetTokenResponse(User user) {
+    public AppleSignInResponse generateTokenAndGetTokenResponse(User user) {
         String accessToken = jwtTokenProvider.generateAccessToken(user.getEmail());
         String refreshToken = jwtTokenProvider.generateRefreshToken(user.getEmail());
-        return AppleTokenResponse.builder()
+        return AppleSignInResponse.builder()
                 .access(accessToken)
                 .refresh(refreshToken)
                 .build();
